Refactor js and html, add css
Changed class name match from equal to regex, because when more class is
added to a div, the match expects all of the class names (not just the
'btn-rock' for example). Now it just has to contain the keyword.

Changed the buttons to divs, added a little style to them.